+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+2.0
+===
+
+* Updated the script to work with GIMP 3.0.
+* Multiple layers can now be aligned at once.
+* Layers can now be aligned against the image instead of another reference layer.
+* Changed the script name to "Align Layers" since the script can now be applied to multiple layers at once.
+
+
 1.0
 ===